+++ src/templates/product-landing.js
@@ -2,11 +2,38 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 import { palette } from '@leafygreen-ui/palette';
+import { cx, css } from '@leafygreen-ui/emotion';
 import { theme } from '../theme/docsTheme.js';
 import { findKeyValuePair } from '../utils/find-key-value-pair.js';
 import useSnootyMetadata from '../utils/use-snooty-metadata.js';
+import NewFeedbackWidget from '../components/Widgets/FeedbackWidget';
 
 export const CONTENT_MAX_WIDTH = 1200;
+
+const formstyle = css`
+  position: absolute;
+  left: 0;
+  bottom: 0;
+  margin-top: ${theme.size.tiny};
+
+  @media ${theme.screenSize.upToLarge} {
+    position: fixed;
+    top: 0;
+    bottom: 0;
+    left: 0;
+    right: 0;
+    overflow-y: auto;
+  }
+`;
+
+const hrStyling = css`
+  border-color: ${palette.gray.light2};
+  margin: ${theme.size.medium} ${theme.size.xlarge};
+`;
+
+const ratingStlying = css`
+  margin: 0px ${theme.size.xlarge};
+`;
 
 const Wrapper = styled('main')`
   ${({ isGuides }) => !isGuides && `margin: 0 auto ${theme.size.xlarge} auto;`}
@@ -220,6 +247,10 @@ const ProductLanding = ({ children, data: { page }, offlineBanner }) => {
     >
       {offlineBanner}
       {children}
+      <hr className={cx(hrStyling)} />
+      <div className={cx(ratingStlying)}>
+        <NewFeedbackWidget slug={'/'} className={formstyle} />
+      </div>
     </Wrapper>
   );
 };
